AN-005 Add JobSorter class
This commit adds the JobSorter class. The purpose of this class is
store all the methods used to sort the jobs between the employees. The
MAXIMISE_PROFIT static method should fulfill the requirements of the
task.

The Employee Class was added to instantiate each employee. Having
each employee as a class will enable convenience methods to be added
to each Employee in the future. Eg promoteEmployee() which could
increase the employee's commission

The sort method on the sortingStrategy also takes in the employees as
an array of Employee to enable the methods to be used if there are
more than 2 employees.